REFACTOR orders button
